refactor: :art: Cambio en la arquitectura del proyecto
