+++ backend/src/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
-import { UserModule } from '../user/user.module';
+import { UserModule } from '../modules/user/user.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 @Module({
