+++ backend/src/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { UserService } from '../user/user.service';
+import { UserService } from '../modules/user/user.service';
 import * as bcrypt from 'bcrypt';
 
 @Injectable()
