+++ backend/src/core/types/pagination-result.ts
