+++ backend/src/modules/items/entities/item.entity.ts
