+++ backend/src/modules/items/items.service.spec.ts
